Remove Fellini
