--- features/support/fellini/abilities/browse_the_web.rb
@@ -1,11 +1,0 @@
-module Fellini
-  module Abilities
-    class BrowseTheWeb
-      include Capybara::DSL
-
-      def self.as(actor)
-        actor.ability_to(self)
-      end
-    end
-  end
-end
